fix(demo): demo sites menu section, demo refactor
